# api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API voyage. Documentation",
        "description": "Cette API permet aux tours-opérateurs de gérer les avis de leurs clients sur les voyages qu'ils proposent.\r\n\r\n    ",
        "contact": {
            "name": "the developer",
            "email": "sengerromain75@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/php/agence_voyage/api/",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/voyage/create": {
            "post": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Ajouter un voyage",
                "description": "Ajouter un voyage",
                "operationId": "a71f7e57c1cc0f0c5825049bae08ad1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "description"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Titre du voyage"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du voyage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inserer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
        "/voyage/read/{voyageID}": {
            "get": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Voir un voyage",
                "description": "Afficher un voyage",
                "operationId": "e21ef168fc52b223ff10d7a30dda4f01",
                "parameters": [
                    {
                        "name": "voyageID",
                        "in": "path",
                        "description": "voyageID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voyage trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Voyage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voyage non-trouvé"
                    },
                    "405": {
                        "description": "Méthode non-autorisé"
                    }
                }
            }
        },
        "/voyage/readAll": {
            "get": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Afficher tous les voyages",
                "description": "Afficher tous les voyage",
                "operationId": "c101af245e85e23833efbf867c8c2ffc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Voyage"
                                    }
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Mauvaise Méthode"
                    }
                }
            }
        },
        "/voyage/update": {
            "put": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Modifier un voyage",
                "description": "Modifier un voyage",
                "operationId": "903d5d855dc7f01e6618bffc82cd8db9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "voyageID",
                                    "titre",
                                    "description"
                                ],
                                "properties": {
                                    "voyageID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titre": {
                                        "type": "string",
                                        "example": "Titre du voyage"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du voyage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inserer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
        "/voyage/delete": {
            "delete": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Supprimer un voyage",
                "description": "Supprimer un voyage",
                "operationId": "f0430e289912b7e5962abda8e699ef07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "voyageID"
                                ],
                                "properties": {
                                    "voyageID": {
                                        "description": "ID du voyage à supprimer",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supprimer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Voyage": {
                "description": "Représente un voyage",
                "properties": {
                    "voyageID": {
                        "description": "Clé unique du voyage",
                        "type": "integer"
                    },
                    "titre": {
                        "description": "Titre du voyage",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description du voyage",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}